sources/telegram: improved code layout
